[Build] Fix cross-platform CI build issues
Windows ARM64:
- Fix size_t to int32_t conversion warning in a64_code_cache_win.cc
- Fix uint64_t to uint32_t conversion warning in a64_backend.cc

Linux ARM64:
- Add Qt system include paths when QT_DIR is not set
- Always link Qt6 libraries on Linux
- Add QtCore/QtGui/QtWidgets subdirectory includes

Linux x86_64:
- Fix discord-rpc cmake compatibility by using os.istarget() for
  platform-specific file inclusion
- Fix FFmpeg cmake compatibility to include x86 architecture files

General:
- Restructure Qt include handling in premake5.lua to properly handle
  both aqtinstall and system Qt packages
- Update FFmpeg and discord-rpc premake files for better cmake support

+++ src/xenia/cpu/backend/a64/a64_code_cache_win.cc
@@ -282,7 +282,8 @@ void Win32A64CodeCache::InitializeUnwindEntry(
   // LDP(X29, X30, SP, POST_INDEXED, 16);
 
   // These opcodes must undo the epilog and put the return address within lr
-  unwind_info->UnwindCodes[0] = OpAllocL(func_info.stack_size);
+  unwind_info->UnwindCodes[0] =
+      OpAllocL(static_cast<int32_t>(func_info.stack_size));
   unwind_info->UnwindCodes[1] =
       UnwindOpWord(UWOP_SET_FP, OpSaveFpLrX(-16), UWOP_END);
